popfloat
Summary: The popfloat library allows the emulation of custom floating point formats on the IPU. The user will be able to specify the float format's mantissa size, exponent size, exponent bias, and also enable or disable denormals as well as Inf/Nan signalling. Once the format has been defined, the user can cast and input (float/half) tensor to the new format and can choose from a range of rounding modes (deterministic or stochastic) used for quantisation. Supported deterministic rounding modes are: Round-to-zero, round to nearest with ties to nearest even, round to nearest with ties away from zero, round towards positive infinity (ceil) and round towards negative infinity (floor). In the case of stochastic rounding, the density function of the noise samples that are used as well as the number of SR bits can be specified. The library also allows the quantised float format to be saved as 8-bit or 16-bit values, provided that float format size (1+mantissa+exponent) can be represented as an 8- or a 16-bit value. Saving and loading to/from a higher precision float format (half/float) is done with the pack and unpack functions.

Reviewers: timh

Reviewed By: timh

Differential Revision: https://phabricator.sourcevertex.net/D12573