Provide an option to make softMax have a scaled output, modify cross entropy loss to accept a scaled output and a scale tensor for its output
Summary:
To improve accuracy of end results and intermediate values especially when the data is of type half.
Added an optional "scaled stable softmax" which improves accuracy of the result by scaling intermediate exponents and the output.
Changed crossEntropyLoss to take a scaling tensor.
Changes to assembler implementations and tests.

Test Plan: Added to/modified lossTest.cpp, nonLinearityTest.cpp

Reviewers: tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D12108