correct mapping of multiUpdated/multiSliced output tensors
Summary:
The output should now be on the same tiles as the slicing vertices, previously
they could be on the wrong tiles.

Test Plan:
updated GatherTest to check that the expected number of elements are present
on the expected number of tiles.
Maually checked this is sane and the mapped regions look correct for the new
GatherTestCase15

Reviewers: jakeh

Reviewed By: jakeh

Differential Revision: https://phabricator.sourcevertex.net/D12647