Improve tile allocation for argMinOrMax (calcAccuracy)
Summary:
The changes are:

1) **popnnCodelets.cpp, popnnCycleEstimators.cpp ReduceClassGather.S, ReduceMaxClassGather.cpp**
Modify the supervisor codelet 'ReduceMaxClassGather' to take a size for the worker data ('divisor') that is not forced to be a power of two. This constraint maybe was a leftover from a previous version and not really needed; in the assembly you just needed to replace a few SHL with MUL.
Also replace the term 'divisor' with 'workerSize' which seems more descriptive.
The original 'divisorlog2' input field in the vertex data was an unsigned short. I've set the new 'workerSize' to be 32 bit to make sure we can accommodate very big row sizes (more than 65535*6 classes ). This is overkill as such row would not fit in the current architecture, but the 2 byte increase in vertex data is largely compensated by the reduction in the number of vertices.

2) **Loss.cpp**

The first reduction stage of argMinOrMax is no longer partitioning the work in chunks of fixed size (32 elements).
It will instead try to give each tile the same number of elements to process.
One or more vertices will be allocated on each tile because each vertex can only process data from a single row (no "2D" vertices have been added).
This is basically the partitioning done by Util.cpp/splitRegions() when 'maxPartitions' is 'tilesPerIPU'

Example:
input is 3x1000; IPU has 4 tiles

we get 6 vertices on the 4 tiles (2 of tiles get one vertex and 2 get two vertices), processing the following elements from each row:

```
  row 0:  [ --------- 750 on tile 0 --------- ]  [250 on tile 1]
  row 1:  [ ---- 500 on tile 1 --- ]  [ ---- 500 on tile 2 --- ]
  row 2:  [250 on tile 2]  [ --------- 750 on tile 3 --------- ]

```

The second and following stages are still partitioning the work in chunks of 32 partial result per (worker) vertex.
The differences fro these stages are:
We now need to keep into account that the rows will have possibly different
number of partials at each stage.
We no longer force all the last stage vertices to reside on the 'result' tile.

A summary of memory/cycle saving before (left) and after (right) the change, for an input of 1277 batches x 8999 classes:

{F509422}

Test Plan:
A custom test tool has been used to test for various sizes of input data.
Maybe I could add the tool to** poplibs/tests**?
Otherwise there some limited testing in poplibs and enigma regressions

Reviewers: godfrey.da.costa, tomm

Reviewed By: tomm

Maniphest Tasks: T9336

Differential Revision: https://phabricator.sourcevertex.net/D12302