Handle serial splits when mapping activations/weights.
Summary:
Ref T9401.

We also set up some of the framework for passing around split ConvIndices because this was necessary to have the mapping work in this new way.

Verified no change in graph construction time/resulting graph for existing resnet 50 batch size 4 training (no serial splits).
Added a test to check the one thing we expect of the mapping when doing a serial split.

Reviewers: richard, samk

Reviewed By: samk

Maniphest Tasks: T9401

Differential Revision: https://phabricator.sourcevertex.net/D12762