Add cycle estimates for looping, dynamic slice and update to the convolution planner
Summary:
Ref T1911. Ref T9403.

When splitting the convolutions serially we will perform the following loop:
```Repeat()
  DynamicSlice()
  Exchange()
  Convolution()
  Reduce()
  DynamicUpdate()```

These change updates the convolution planner to take into the costs of slicing and updating as well as performing everything N times (depending on how big a serial split we decide to do).

When we are not splitting serially the cost is the same as it was before (both the slice and update estimates return 0 in that case).

Test Plan: existing tests didn't fail -- hard to test the correctness accuracy of the estimate functions.

Reviewers: tomm, richard, godfrey.da.costa, davel

Reviewed By: richard

Maniphest Tasks: T9403, T1911

Differential Revision: https://phabricator.sourcevertex.net/D12212