Fixed a bug in the convolution planner where it would unconditionally prune for cycles
Summary:
Depends on D12553, D12779. Fixes T9644.

Before the model would always prune the search space based on the best cycle cost found so far. This is incorrect when we want to minimise for memory as it would cause lower-memory plans to not be found.

Additionally, we never pruned the search space based on the best memory cost found so far which would cause a compile time cost, which has been fixed too. Another memory issue was that we did not take into account the fact that the memory of the transformations could be greater than the convolutions -- this case appeared as a result of the changes in this diff and so therefore was fixed here too.

For ResNet50, batch size 4, training this change has the following effect:

|                    | Total Memory  | Max Tile Memory | Always Live          | Max Live Bytes
| ------------  | ----------------  | --------------------  | -------------------- | ------------------
| Before        | 342,376,084   | 306,698                 | 182,199,845       | 338,450,981
| After           | 337,826,590   | 303,366                 | 192,237,744       | 326,254,768
| Difference | 4,549,494        | 3,332                     | **-10,037,899**  | 12,196,213

The always live memory increase is roughly split across:
 - Exchange code (4MB)
 - Vertex and control code (2MB)
 - Vertex data (4MB)
As the planner does not consider always live memory this is unfortunate but hard to avoid.

Test Plan: `./test.sh poplibs -R conv`, additionally verified manually with the gemm tool that the smallest plan is now being found.

Reviewers: tomm, richard

Reviewed By: tomm, richard

Maniphest Tasks: T9644

Differential Revision: https://phabricator.sourcevertex.net/D12656