Masking of loop variable to allow the compiler to use rpt, with protection in elementwise ops.
Summary:
Mask out the loop vaiable in C dispatch methods, which allows the compiler to use the rpt counter.  See T9902 for details of what was necessary to make this happen
This brings up the possibility that the data is larger than the rpt counter limit (this
was actually already true in some cases before), although in practice with:

1. A limited amount of memory.
2. 2 floats or 4 halves per loop.
3. The dispatch methods only being used for float, half based loops.
4. Further work division for a supervisor.
5. Multiple operands in some cases, of the same size which use up the limited memory.

It is unlikely that the repeat limit will be hit. So a simple mechanism is used - if we have one
region and it is too big to divide up between workers, use a 2D vertex instead.  Then, split
any regions so that we don't exceed the limit.  A simple method of work division is used.

Fixes T7045

Test Plan: Existing tests (StdOperatorsTest), verified using a smaller limit manually. Addition of tests to LargeRegionsSplitTest.cpp.

Reviewers: tomm

Reviewed By: tomm

Maniphest Tasks: T7045

Differential Revision: https://phabricator.sourcevertex.net/D12662