Add serial output channel splitting to the planner
Summary:
Ref T9403.

This change adds support for splitting convolutions serially in the output channel dimension. This is only supported for single-IPU convolutions for now.

These are disabled by default (until T9401 is ready). Because of this this change has no effect on any of the rFRAMEWORKCOMP benchmarks (only tested on the enigma ones).

Test Plan: manually tested, added a unit test for sanity check as this code is not hit by default

Reviewers: tomm, richard

Reviewed By: richard

Maniphest Tasks: T9403

Differential Revision: https://phabricator.sourcevertex.net/D12284