Make calculation of softmax robust when scaling is used
Summary:
The error between the max of the softmax input and the log(SOFTMAX_SCALING)
can be such that exp(actual) * exp(error) exceeds half. SOFTMAX_SCALING
was reduced to counter exactly that problem but the values that softmax
input can be such that the scale has to be reduced further which is
detrimental to maximising precision. Instead this change first subtracts
off the max and then applies the scaling.

Reviewers: tomm, timp

Reviewed By: timp

Differential Revision: https://phabricator.sourcevertex.net/D12921