Copy tensor after cloning for non-inplace identity map expression
Summary:
When requested to do a non-inplace identity operation in map() (i.e. the expression passed to map is just a placeholder) we were cloning the tensor but not copying it.
With this diff we just add the Copy program.

This addresses T9425 popops::map returns zero tensor when the given expression is a placeholder

Test Plan:
create a test program containing
`auto out = map(g, popops::expr::_1, {in}, prog);`
and verify that `out` gets the same elements as `in`

Reviewers: godfrey.da.costa, timp

Reviewed By: godfrey.da.costa

Maniphest Tasks: T9425

Differential Revision: https://phabricator.sourcevertex.net/D13014