Fix to size of reduction length fields for CPU implementation
Summary:
Bug fix - see T9764.  But basically, reductions with a large tensor input can produce vertices with count/partials sizes greater than that supported by an unsigned short in the vertex state.  This isn't going to happen on IPU, given memory constraints, and work division between workers.  But it can happen on the CPU (all data on one tile) - so use of a larger data type on the CPU was the solution used.

Fixes T9764

Reviewers: samuelh

Reviewed By: samuelh

Subscribers: O1 Reduction

Maniphest Tasks: T9764

Differential Revision: https://phabricator.sourcevertex.net/D12996