Add estimates to account for the zero padding added by the AMP vertices
Summary:
Depends on D12927. Fixes T9834.

This change adds cycle and memory estimates for the zero padding that is added when the kernel size is not a multiple of the weights per conv unit.

This change blocks landing D12935 and by itself causes a significant regression for the `resnet18_bs8_gn` benchmark (+70MB) -- however with both this diff and that diff the benchmarks show a large improvement:

|        benchmark         | difference in total memory | difference in max tile memory | difference in cycles |
|--------------------------|----------------------------|-------------------------------|----------------------|
| conv_2_tr                |                   -3997712 |                             0 |               -11354 |
| lstm_bs1_seq32_3x256x256 |                          0 |                             0 |                    0 |
| lstm_bs1_seq5_28x256     |                    3183168 |                          -512 |                10267 |
| matmul3_tr               |                          0 |                             0 |                    0 |
| resnet18_bs1_bn          |                   -1511074 |                             0 |                17735 |
| resnet18_bs1_gn          |                   -1528162 |                             0 |                17735 |
| resnet18_bs4_bn          |                  -48072290 |                        -32816 |                -8326 |
| resnet18_bs8_gn          |                  -29305956 |                         -4256 |                21265 |
| resnet50_bs1_bn          |                     499312 |                           384 |                62653 |
| resnet50_bs1_bn_inf      |                     456350 |                             0 |                 4415 |
| resnet50_bs1_gn          |                     501307 |                            96 |                69609 |
| resnet50_bs1_gn_recomp   |                     743259 |                          -608 |                74468 |
| resnet50_bs2_gn          |                    -247408 |                         -5768 |               319944 |
| resnet50_bs2_gn_recomp   |                     -35367 |                         -4032 |               420469 |
| resnet50_bs4_gn          |                   -2047304 |                          7196 |               276057 |
| resnet50_bs4_gn_recomp   |                   -1861256 |                          5532 |               344937 |

A similar pattern exists if D12935 is landed without this diff therefore these two commits should be seen as being tightly coupled.

The 3MB increase in the `lstm_bs1_seq5_28x256` benchmark is a result of a 4MB increase in vertex code -- as this is a tiny model this always live memory will be amortized as the model gets larger. Also the peak temp memory and max tile memory both went down as a result of this change so that benchmark actually shows an improvement too, despite the total memory increase.

Test Plan: `./test.sh poplibs` and the benchmarks

Reviewers: tomm, richard

Reviewed By: richard

Maniphest Tasks: T9834

Differential Revision: https://phabricator.sourcevertex.net/D12934