Add support for GELU non-linearity in poplibs
Summary:
GELU (Gaussian Error Linear Unit) is more widely used in
NLP (and other ML models) is a smoothed RELU function and
is continuous throughout its domain.

This implementation uses an approximation of the CDF of
a normal distribution which is
   Phi(x) = (1 + tanh( x * alpha * (1 + beta * x * x)))

This nonlinearity differs from others already implemented
in that it takes the input of the non-linearity rather than
is output to compute the gradient.

Only a C++ vertex is implemented with placeholder values
for cycles estimates.

Test Plan: Added test

Reviewers: tomm

Reviewed By: tomm

Maniphest Tasks: T10615

Differential Revision: https://phabricator.sourcevertex.net/D13696