Add functions to Poplibs for checking for the existence of NaNs and infinities
Summary:
Fixes T10149. This diff is also the basis for my symposium talk.

Adds a new function: `hasNaN` which returns a scalar boolean value if any of the input is a NaN. It also adds support for `IS_NAN` and `IS_INF` to the popops map expressions, these return an output of the same size with each element mapped to whether or not the same element in the input is a nan/inf.

Note that this does not include support for these two ops in expression fusions.

Test Plan: new tests

Reviewers: timp, oleksiik, claudiop, tomm, timh

Reviewed By: oleksiik

Subscribers: chadj, Eddy, timh

Maniphest Tasks: T10149

Differential Revision: https://phabricator.sourcevertex.net/D13089