Don't use getInitialValue to get single elements
Summary: Previously this used getInitialValue() to get single elements of a 1D field. This was a slightly weird way to do it in this case, since it got every element that way and you can just use getInitialValues() to do that. Also I am going to remove the ability for getInitialValue() to get single elements of a 1D field, because it is a bit weird that you can do it - you can't get single elements of a 2D field for example.

Test Plan: CI

Reviewers: timp

Reviewed By: timp

Differential Revision: https://phabricator.sourcevertex.net/D13916