CastSupervisor vertices
Summary:
This add Cast supervisor vertices for float and half, to be used when there is a single contiguous region on a tile.
Using the Supervisor vertex adds 20 extra cycles (for the workers)

Test Plan: Added a --supervisor option to the CastTest test executable (part of tests/lib/popops/codelets) and a modified a couple of test cases in CmakeLists.txt to make use of it.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Subscribers: davel

Maniphest Tasks: T7348, T4817

Differential Revision: https://phabricator.sourcevertex.net/D13404