Fold independent spatial dimensions into channels if there are very few in pooling layers.
Summary:
This is limited to just when no. of channels is < channel grain size to keep this targeted.
This should always be better than not doing this as when channels are less than the grain size we have to add padding when we may as well fold in other independent dimensions so we're actually doing some work in these other channels.

Reviewers: samk, godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D13951