Assembly versions of reduce partials equal size, implementing half out variants and update variants.
Summary:
Fairly simple assembly changes using macros to implement more of the variants of this vertex. We should now have all
variants where the operation is Add or SquaredAdd implemented in assembler (combinations of scaled=true/false, update=true/false, input=float/half, output=float/half)
The function for selecting which is assembly, and which is C could be simplified, but I imagine adding more variants - max, min etc in future and it may be useful there too.

Reviewers: samk

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D14352