Handle continuous reductions with suitable vertex
Summary:
Before when the inputs and output regions of the tensor were continuous
the reductions forced a copy to spread the data and then to reduce kept a pointer to
every element in the tensor. This change avoids both of these by having a vertex that handles
continuous inputs. This could also mean we can delete the single input/output region specialisation based on:
```
The new vertex uses:
One large partials edge, which is a contiguous block of memory, divided into numOutputs groups of numPartials elements
One output edge, of size numOutputs.
The output is treated as a series of scalars, where numPartials consecutive elements reduce to a single output scalar.

Scale and update are supported (all versions in assembler)
Assembler exists for float, floatOut / half, floatOut / float, halfOut and half, halfOut versions of add, squareAdd
```

```
Specialisation 2 (scalar output single input) uses:
One partials edge (size = numPartials)
One scalar output
The partials edge is simply reduced to the single scalar output.

Update is supported but not in assembly
Scale is not supported anywhere
Assembler exists for float, floatOut and  float, HalfOut versions of add, squareAdd

```
... although the vertex state of Specialisation 2 is more compact, so I'd leave a tidy up of this for another time maybe?

Reviewers: samuelh, tomm, samk

Reviewed By: samk

Subscribers: timp, O1 Reduction

Differential Revision: https://phabricator.sourcevertex.net/D13868