Add VarStructure utility to create and handle tensors with a certain partitioning.
Summary:
Two new utility functions:
* addVariableWithSplits - add a variable where each dimension is split a certain number of times.
* iterateVariableSplits - given a tensor and how it was split, iterate the slices of that tensor.

The meaning here of 'tensor with a certain partitioning' is that each partition is contiguous in memory when introspected.
This is a generalised version of the way dynamic slicing maps its tensors such that each tile has a certain portion of the whole tensor to slice but when introspected individually each tile's slice of the tensor is contiguous.

Reviewers: Eddy

Reviewed By: Eddy

Subscribers: samk

Differential Revision: https://phabricator.sourcevertex.net/D14805