In scaledAddTo variants first cast all tensors to be of result type
Summary:
The general form of the scaledAddTo operation is

x = a * x + b * y

If the types for a, x and y are different, then the
exact arithmetic used in the operation has to be
specified. Until this is done to cater to different
use cases: a, b and y are first cast to be of type x,
and then the operation performed.

This means that the scaledAddSupervisor vertices which
have y of type float and x of type half to be never used.
But these scaledAdd didn't have a 2D version anyway.

T11658

Test Plan: Modified test

Reviewers: timp

Reviewed By: timp

Subscribers: ranjayc

Differential Revision: https://phabricator.sourcevertex.net/D15274