Adding a better test coverage for ReducePartialsEqualSize and developing max,min assembler
Summary:
ReducePartialsEqualSize reduction vertex didn't have a specific test, so this diff adds one.  No bugs were found in the original code, having improved the testing.

Then, assembler versions for max, min were added - the testing was useful for these, and the reduction connection logic changed to select this vertex where allowed - as the assembly versions now exist and the implementation should be more efficient.

Note - I did implement all assembly in 1 file (1st diff version here includes a similar structure).  This works but was a bit of a complex use of conditional assembly, so changed to 2 files.  In doing that 1 file version I used the .ifeqs and .ifnes assembly directives.  They behave wrongly if nested (In the macro in the place I'm now using .ifc).  The effect was like when nested, all conditions would appear as true!

Reviewers: samuelh, samk

Reviewed By: samk

Subscribers: O1 Reduction

Differential Revision: https://phabricator.sourcevertex.net/D15245