Adding scaledAdd half, half, float variant, changing casting in codelets and adding tests
Summary:
Addition of a new scaledAdd with AType = half BType = half and scaleType = float.
The reason for this is that with small scale, the scale could be quantised out when treated as a half.  Storing/using it as a float would be a solution - casting to happen in the codelet. (A similar issue with a large scale is also solved as a by product).
Added tests to check that it functions (But not to verify that it solves the quantisation issue)
Assembler version to follow, along with logic to select it in scaledAdd.cpp.

Presently this should have no effect other than the tests, as scaledAdd.cpp will not create these codelets.

Reviewers: samk, samuelh

Reviewed By: samk

Subscribers: ranjayc

Differential Revision: https://phabricator.sourcevertex.net/D15499