popops function to check the accuracy of a scalar float Tensor when cast to a half, integrated into the scaledAdd functions with codelets in C++
Summary:
Added a helper codelet (In C only at the moment) to check error when casting a scalar tensor.  API supports all types, but internally only float to half, half to float are supported. In the float to half version , if the value is in the denorm range the error is compared to a tolerance and a resulting unsigned Tensor is returned to indicate if the error is too great.  The cast result is also return, as it is highly likely that it will be used in the case that its accuracy is good.

Added a test to check that a codelet/codelet that implement scaled Add for half data, float scale have improved accuracy.
Added graph creation time check for cast accuracy, which is used when deciding if a constant can be used when cast at compile time.

(Includes the work from D15426 and D15499.  Plus changes to the test and the scaledAdd.cpp function to select appropriate vertices as required)

Fixes T11937

Reviewers: tomm

Reviewed By: tomm

Subscribers: ranjayc

Maniphest Tasks: T11937

Differential Revision: https://phabricator.sourcevertex.net/D15554