Pattern detection for reductions.
Summary:
Change to the way that reduction partials are listed in each tile level RegionReduction structure.  The method used is to look at each reduction and each contiguous memory area for that reduction and identify which data is needed for that reduction.  As we are always viewing this as reducing down a column of a rank 2 tensor this means identifying the data that is in a specific column.

Having scanned any contiguous memory area for data that is needed by the reduction this can be represented by a pattern with:
start: offset to the pattern start
length: number of consecutive tensor elements belonging to the reduction
stride: length of a repeating pattern if one was identified.
repetitions: number of repeats of the pattern identified.

There can be multiple patterns (per reduction) in any memory area, and multiple memory areas on tile, pieces of which are required for any reduction to be completed.

This work allows the plan for on-tile reductions to be made, identifies patterns and uses the patterns to determine the partials.  The end effect should be the same as before, however using patterns maintains more information about the tensor layout than the previous method.  We should eventually be able to use patterns to target new vertices that assume data lies in a single contiguous memory area (although is not itself contiguous).

Reviewers: tomm

Reviewed By: tomm

Subscribers: samuelh, O1 Reduction

Differential Revision: https://phabricator.sourcevertex.net/D13728