Asin operator added to popops expressions
Summary:
I added the ASIN operator to popops available expression in order to be able to add the same operator in popart.
Here is the popart revision https://phabricator.sourcevertex.net/D15721 if you need it.

Test Plan:
I tested the correct behaviour of the popart operator asin, which is behaved as expected.
I added the asin test to the test list and added it to the CMakeList.txt

Reviewers: samk, tomm

Reviewed By: samk

Maniphest Tasks: T6437

Differential Revision: https://phabricator.sourcevertex.net/D15829