Add createOutputForElementWiseOp and use it for element-wise ops in poplibs
Summary:
Ref T12259.

This is as opposed to mapOutputForElementWiseOp which isn't really the right solution. Much of this is copied from that implementation.

We now clone the input selected as the best candidate (most well spread, no aliases/constants) and when no input was selected for cloning we just create a new variable with the same shape and map linearly with the defaults rather than minGrainsPerTile/grainSize which weren't used and defaulted to 1/0 which resulted in tensors being too thinly spread often.

Frameworks comparisons enigma benchmark show no regression or a minor improvement in memory usage.

Reviewers: Eddy

Reviewed By: Eddy

Maniphest Tasks: T12259

Differential Revision: https://phabricator.sourcevertex.net/D16146