Doing float arithmetic which converting invStdDev(half) to variance(half) and converting invStdDev(half) to/from variance(float)
Summary:
Firstly - adding codelets to use float for the internal arithmetic when converting invStdDev to variance when both are of type half.  This would avoid the potential overflow in `(1/(in*in)) - epsilon` when  a large `in` is squared, but can produce a valid half output.

Second - a new API to allow  invStdDev(half) to be converted to variance(float) - all arithmetic in float
variance(float) to be converted to invStdDev(half) - all arithmetic in float.
This doesn't us the map function and cannot be done inPlace. Use:
`invStdDevToVariance(Graph &graph, const Tensor &src, const Tensor &epsilon, Sequence &prog, const Type dstType, const std::string &debugPrefix)`
This will function if src.elementType() ==dstTyp, or dstType = HALF for invStdDev out, or dstType = FLOAT for variance out.

Also, noted that the existence of the `BroadcastScalar2D` and `BroadcastScalar2DInPlace`  codelets for variance conversions is odd. These expect to see a 2D vector [sizex][sizey], and a B vector[sizex].  This is useful when finding patterns of aliased data on tile when doing mpy,add,sub operations but not for variance conversions where we expect  epsilon (a B scalar) to be a scalar only.  The codelets existing wasn't a big deal when we had a templated C implementation, but now adding assembler it is.  The solution is to remove them completely and error if the epsilon/B is of the wrong size.

Fixes T12134

Reviewers: tomm, samk, samuelh

Reviewed By: samk

Maniphest Tasks: T12134

Differential Revision: https://phabricator.sourcevertex.net/D16002