Decision to use half or float scale in the scaledd add<half,half,float> vertex to avoid huge vertex state.
Summary:
Instead of using an if or a switch statement we make the decision to take the (float) scale and use it directly as a float in a scaledAdd<half,half,float> vertex or to cast it to a half and use a scaledAdd<half,half,half> vertex.  The decision is made in the scaledAdd<half,half,float> vertex directly meaning that there is one vertex state, not two.

Presently the decision variable is a bool, and the decision function returns a tensor with the scale cast to half.  The tensor looks redundant, and is bool the best choice for the variable given the difficulty in writing a bool?

Reviewers: samuelh

Reviewed By: samuelh

Differential Revision: https://phabricator.sourcevertex.net/D16335