Don't unroll loop in collectives
Summary:
Use a repeat for collective operations instead of unrolling the for loop. Forces the non single image version to use dynamic slice.

This changes the unidirectionalCollectivePrograms to create a CollectivesProgram structs. The bidirectional and meet in the middle methods
call the unidirectional methods to get 2 of these structs. THere are 3 more functions provided for converting these structs into a sequence for
each collective method (ring, bidirectional, meet-in-the-middle)

Reviewers: richard

Reviewed By: richard

Maniphest Tasks: T12049

Differential Revision: https://phabricator.sourcevertex.net/D16166