Infer the tile mapping for constants from the rest of the expression
Summary:
Previously constants were always placed on tile 0, now we try to place
them on a tile that one of the variables is mapped to.

The motivation for this change is that Poplar's subgraph replication pass
only replicates subgraphs containing scalar variables which are mapped to
the same tile so placing the constant on a different tile can inhibit
subgraph replication.

Reviewers: timp, godfrey.da.costa

Reviewed By: timp, godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D16541