Don't explicitly transpose/flipxy convolution weights in matmul weight update pass to avoid transposing unnecessarily.
Summary:
There is no xy dimension in the conv weights for the matmul in the weight update pass and introspection will deal with input/output channels being transposed if necessary depending on the plan.

This is motivated by a case where the operands were swapped in the plan meaning we explicitly transposed input channels and output channels and then transposed them back due to swapped operands in the weight update pass.

Reviewers: richard

Reviewed By: richard

Subscribers: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D16582