Improvements to building GP objects
Summary:
Factored out the cmake code used to create GP files in poplibs. Ref T12253.

This refactor includes a number of improvements to decrease the build time:
 - Automatically build all cpp source files separately (and therefore in parallel)
 - Don't build the _c versions unless we are planning to run the Sim:cpp tests
 - Don't build for the ipu0 target
 - Don't add a dependency on the gp files for the associated library to be built

On top of this we can still build each target in parallel too but there is a bug in Poplar that prevents us from merging GP files with different targets. Once this is fixed that change will be added in a separate diff.

Test Plan: `./test.sh poplibs` and examined the build logs when building the library.

Reviewers: richard, Eddy, samuelh, tomm, claudiop

Reviewed By: richard

Maniphest Tasks: T12253

Differential Revision: https://phabricator.sourcevertex.net/D16262