Added a combineConvGroups transformation
Summary:
Fixes T9168.

Added a new transformation to the planner that, for example when we have a depthwise convolution (i.e. each conv group has one input and output channel), we would transform this into a grouped convolution with N/4 groups each with 4 input and 4 output channels. This expands the weights by a factor of 4, but we don't need to pad the activations (which is what results in the expensive copies).

For the first pass this transformation will only handle a single input channel but can be extended to handle pairs of input channels too.

Running the mobile net that is found in rPUBLICEXAMPLES i get the following numbers:

|              |    before     |    after     |
|--------------|---------------|--------------|
| memory       | 118,563,602 B | 84,652,377 B |
| cycles       | 924,552,762   | 440,244,762  |
| compile time | 141 sec       | 61 sec       |

Test Plan: new tests

Reviewers: richard, tomm, Richardc, godfrey.da.costa, Eddy

Reviewed By: richard

Subscribers: manueles

Maniphest Tasks: T9168

Differential Revision: https://phabricator.sourcevertex.net/D15544