Use poplar::VertexBase for all supervisor codelets that are only partially implemented in assembly
Summary:
Ref T11040.

This change means that we never need to run a poplibs SupervisorVertex as a pseudo supervisor, if there is only a C++ compute method then the codelet becomes a worker vertex.

The elementwise 1D supervisor ops that are now conditionally supervisor should be renamed but that was quite involved due to assembly, macro and name clashing with the compute component of each vertex that it was left out of this change.

Test Plan: `./test.sh poplibs`

Reviewers: richard, Eddy, timp

Reviewed By: richard

Subscribers: davidb, leandrov

Maniphest Tasks: T11040

Differential Revision: https://phabricator.sourcevertex.net/D17087