Split work respecting grain size when a continuous reduce is expected
Summary:
Deal with the case that continuousReduce and potentially reductions using specialisation 2 should be split respecting the grain size of the type of data they are reducing.  This is because they differ from other reductions, which make a parallel sum of a number of columns, eg `ABCD,ABCD,ABCD` N times.  It is OK in that case to split N into separate worker tasks as it is the length of `ABCD` that affects alignment of the vertex input data, not the work split.

On the other hand continuous reduce and specialisation2 deal with a section of memory: `AAAAAAAAAAAAAAA ...` if we split that and ignore the grainSize the start of work for a given vertex will become misaligned and copies will be required.

Fixes T11259

Reviewers: tomm

Reviewed By: tomm

Maniphest Tasks: T11259

Differential Revision: https://phabricator.sourcevertex.net/D17083