Using pattern detection for reductions to reduce the data rearrangement when columns are non-contiguous on tile
Summary:
Deal with column data for reductions where the column number is non-contiguous but the data is laid out as if in columns in memory.  The reductions will happen without a re-arrangement pre reduction ( which is what happened before).  If we are producing an output tensor with all reduction happening on tile then no re-arrangement will be necessary at all.  Otherwise, if the output is already created re-arrangement will happen after reduction (at a lower cost than before), or if there are intermediate reductions it will happen after the first, on-tile reduction.

This mainly required re-work to get to return the tensor form the point at which the layout of data was discovered.

Some further work to redo the way that the intermediate partials are stored/sorted into column order is required to eliminate the re-arrangement before the first intermediate stage.
Fixes T12587

Test Plan: Regular tests.  Added some variants or reduce_op tests with shuffling of columns.

Reviewers: tomm

Reviewed By: tomm

Subscribers: samuelh, marieanne, O1 Reduction

Maniphest Tasks: T12587

Differential Revision: https://phabricator.sourcevertex.net/D16608