Test tool for generating Broadcast patterns
Summary:
Any broadcast pattern that can be described using the 3 integers (innerFactor, motifLength, outerFactor) can be generated with this test tool.
A description of broadcast patterns is provided within the source file.
The tool assumes that the requested pattern can fit on a single tile.

Fixes T12449

Reviewers: timp

Reviewed By: timp

Maniphest Tasks: T12449

Differential Revision: https://phabricator.sourcevertex.net/D16339